omicscode
adding the support for the sam files. finished coding all the doseqGO in Axum and preparing for the release. Watched one video this morning and coded the entire doseqGO from GO to RUST.

I code RUST so fast that it took me to code in GOLang 3 days whereas in RUST 4 hours, I finished all of this.